Part 9 - Collisions & Textures
This part of the tutorial was really a messy experience. My code is the
exact same as theirs. Literally copied and pasted theirs on top of mine
after typing it. Did not fix my issue. Must be a difference in the scene
I didn't notice but the collision mesh doesn't work well at all.